protect "key" file

+++ initial_setup.sh
@@ -1,3 +1,5 @@
+#!/bin/sh
+
 echo "-------------------------"
 echo "agama lib. | agora nostr "
 echo "-------------------------" 
@@ -8,10 +10,16 @@ pip install -r requirements.txt
 
 export PYTHONPATH="$PYTHONPATH:./"
 
-echo "--- rename ------------------"
-echo "agama_nostr/nostr_key.py.tmp "
-cp ./agama_nostr/nostr_key.py.tmp ./agama_nostr/nostr_key.py
+
+# protection against accidental overwriting of an existing file
+if [! -f "./agama_nostr/nostr_key.py"]; then
+    echo "--- rename ------------------"
+    echo "agama_nostr/nostr_key.py.tmp "
+    cp ./agama_nostr/nostr_key.py.tmp ./agama_nostr/nostr_key.py
+    echo "-> agama_nostr/nostr_key.py"
+fi
 
 echo
-python nostr_key_gen.py
 
+# automatic generation of a random key that can be used
+python nostr_key_gen.py
